+++ sdk/log/logger.go
@@ -56,6 +56,10 @@ func newLogger(p *LoggerProvider, scope instrumentation.Scope) *logger {
 }
 
 func (l *logger) Emit(ctx context.Context, r log.Record) {
+	if l.provider.stopped.Load() {
+		return
+	}
+
 	newRecord := l.newRecord(ctx, r)
 	for _, p := range l.provider.processors {
 		if err := p.OnEmit(ctx, &newRecord); err != nil {
@@ -71,6 +75,10 @@ func (l *logger) Emit(ctx context.Context, r log.Record) {
 // processed, true will be returned by default. A value of false will only be
 // returned if it can be positively verified that no Processor will process.
 func (l *logger) Enabled(ctx context.Context, param log.EnabledParameters) bool {
+	if l.provider.stopped.Load() {
+		return false
+	}
+
 	p := EnabledParameters{
 		InstrumentationScope: l.instrumentationScope,
 		Severity:             param.Severity,